+++ aseda-jvf/src/main/java/se/leiden/asedajvf/service/JwtService.java
@@ -21,7 +21,7 @@ public class JwtService {
         this.secretKey = Keys.secretKeyFor(SignatureAlgorithm.HS256);
     }
 
-    public String createToken(String username, Long userId, Role role) {
+    public String createToken(String username, int userId, Role role) {
         Date now = new Date();
         Date validity = new Date(now.getTime() + validityInMilliseconds);
